+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Add experimental Admin layout (PR #371)
 * Add the [GOV.UK Frontend](https://design-system.service.gov.uk/) library to the gem (PR #398)
+* Allow linking to the Design System on component pages (PR #401)
 
 ## 9.3.6