+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add the [GOV.UK Frontend](https://design-system.service.gov.uk/) library to the gem (PR #398)
 * Allow linking to the Design System on component pages (PR #401)
 * Add govuk:analytics:organisations meta tag if the current page is an organisation (PR #397)
+* Move the highlight box component from Collections to the gem (PR #403)
 
 ## 9.3.6